S O M E, javadoc, smell good ;)
