+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/api/event/ProtectionCheckStatus.java
@@ -19,6 +19,9 @@
 
 package com.ghostchu.quickshop.api.event;
 
+/**
+ * The status about protection check current status.
+ */
 public enum ProtectionCheckStatus {
     BEGIN(0),
     END(1);
