+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/api/event/package-info.java
@@ -18,6 +18,6 @@
  */
 
 /**
- * The QuickShop events that allow addons to listening
+ * The QuickShop events that allow addons to listen to.
  */
 package com.ghostchu.quickshop.api.event;
