+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/command/subcommand/SubCommand_Transfer.java
@@ -21,7 +21,7 @@ package com.ghostchu.quickshop.command.subcommand;
 
 import com.ghostchu.quickshop.QuickShop;
 import com.ghostchu.quickshop.api.command.CommandHandler;
-import com.ghostchu.quickshop.api.event.ShopOwnerShipTransferEvent;
+import com.ghostchu.quickshop.api.event.ShopOwnershipTransferEvent;
 import com.ghostchu.quickshop.api.shop.Shop;
 import com.ghostchu.quickshop.util.Util;
 import org.bukkit.entity.Player;
@@ -73,7 +73,7 @@ public class SubCommand_Transfer implements CommandHandler<Player> {
             }
             List<Shop> shopList = plugin.getShopManager().getPlayerAllShops(fromPlayer.getUniqueId());
             for (Shop shop : shopList) {
-                ShopOwnerShipTransferEvent event = new ShopOwnerShipTransferEvent(shop, shop.getOwner(), targetPlayer.getUniqueId());
+                ShopOwnershipTransferEvent event = new ShopOwnershipTransferEvent(shop, shop.getOwner(), targetPlayer.getUniqueId());
                 if (event.callCancellableEvent()) {
                     continue;
                 }
